+++ src/utils.ts
@@ -54,6 +54,7 @@ import {
   setNewTokensLimitPerAddressFunctionName,
   claimScFundsFunctionName,
   getMintedPerAddressPerDropFunctionName,
+  populateIndexesFunctionName,
 } from './config';
 
 export const baseDir = cwd();
@@ -285,9 +286,7 @@ export const getMintTransaction = (
   } else {
     return contract.call({
       func: new ContractFunction(mintFunctionName),
-      gasLimit: new GasLimit(
-        baseGasLimit + (baseGasLimit / 1.6) * tokensAmount
-      ),
+      gasLimit: new GasLimit(baseGasLimit * tokensAmount),
       args: [new U32Value(tokens)],
       value: Balance.fromString(tokenSellingPrice).times(tokens),
     });
@@ -476,6 +475,7 @@ export const commonScQuery = async ({
   args?: TypedValue[];
 }) => {
   const smartContractAddress = getTheSCAddressFromOutputOrConfig();
+  const spinner = ora('Processing query...');
   try {
     const provider = getProvider();
     await syncProviderConfig(provider);
@@ -490,7 +490,6 @@ export const commonScQuery = async ({
       smartContractAddress
     );
 
-    const spinner = ora('Processing query...');
     spinner.start();
 
     const response = await scQuery(functionName, smartContract, provider, args);
@@ -508,6 +507,7 @@ export const commonScQuery = async ({
     console.log('Query results:');
     console.log(`${resultLabel}: `, result.trim());
   } catch (e) {
+    spinner.stop();
     console.log(e);
   }
 };
@@ -565,5 +565,17 @@ export const getClaimScFundsTransaction = (
   return contract.call({
     func: new ContractFunction(claimScFundsFunctionName),
     gasLimit: new GasLimit(gasLimit),
+  });
+};
+
+export const getPopulateIndexesTx = (
+  contract: SmartContract,
+  gasLimit: number,
+  amount: number
+) => {
+  return contract.call({
+    func: new ContractFunction(populateIndexesFunctionName),
+    gasLimit: new GasLimit(gasLimit),
+    args: [new U32Value(amount)],
   });
 };